create separate irc/connection_limiting package
